Generating a Dish no longer shuffles dishArray
The currently displayed Dish in the generate Dish tab gets removed
from the dishArray when generating or rolling a dish and then readded
to the array to each generate click will show a different dish than the
currently displayed dish. Before this commit, the removed dish was
readded with push which reassambled the order of the array.
now the dish is readded with splice and the indexnumber